backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
bcrypt==4.0.1
python-multipart==0.0.20
openpyxl==3.1.5
openai==1.107.3
